Add IContentManager.GetAsync() extension to allow retrieve multiple content items using ids and version options
